EditTodo Fragment migration and navigation
