+++ app/src/main/res/navigation/app_navigation.xml
@@ -13,10 +13,22 @@
         <action
             android:id="@+id/navigateToCreateTodo"
             app:destination="@id/createTodoFragment" />
+        <action
+            android:id="@+id/navigateToEditTodo"
+            app:destination="@id/editTodoFragment" />
     </fragment>
     <fragment
         android:id="@+id/createTodoFragment"
         android:name="com.zachtib.simpletodo.ui.createtodo.CreateTodoFragment"
         android:label="CreateTodoFragment"
         tools:layout="@layout/create_todo_fragment"/>
+    <fragment
+        android:id="@+id/editTodoFragment"
+        android:name="com.zachtib.simpletodo.ui.edittodo.EditTodoFragment"
+        android:label="EditTodoFragment"
+        tools:layout="@layout/edit_todo_fragment">
+        <argument
+            android:name="todoItemId"
+            app:argType="integer" />
+    </fragment>
 </navigation>
